refactor: changes components directories structure
